build: conditionally copy public in next build

+++ toolchains/workspace-pnpm/build_next_build.py
@@ -34,18 +34,20 @@ if __name__ == "__main__":
         dirs_exist_ok=True,
         )
 
-    shutil.copytree(
-        os.path.join(args.package_dir,"public"),
-        os.path.join(args.out_path, ".next", "standalone", "public"),
-        symlinks=True,
-        dirs_exist_ok=True,
-        )
+    if os.path.exists(os.path.join(args.package_dir, "public")):
+        shutil.copytree(
+            os.path.join(args.package_dir,"public"),
+            os.path.join(args.out_path, ".next", "standalone", "public"),
+            symlinks=True,
+            dirs_exist_ok=True,
+            )
 
-    shutil.copytree(
-        os.path.join(args.package_dir,".next", "static"),
-        os.path.join(args.out_path, ".next", "standalone", "static"),
-        symlinks=True,
-        dirs_exist_ok=True,
-        )
+    if os.path.exists(os.path.join(args.package_dir, ".next", "static")):
+        shutil.copytree(
+            os.path.join(args.package_dir,".next", "static"),
+            os.path.join(args.out_path, ".next", "standalone", "static"),
+            symlinks=True,
+            dirs_exist_ok=True,
+            )
 
     sys.exit(exit_code)
